requirements.txt
beautifulsoup4==4.4.1
bleach==1.4.2
boto==2.38.0
Django==1.8.3
django-admin-honeypot==0.4.0
django-admin-tools==0.6.0
django-appconf==1.0.1
django-betterforms==1.1.2
django-bootstrap-form==3.2
django-compressor==1.5
django-crispy-forms==1.5.2
django-debug-toolbar==1.3.2
django-html-sanitizer==0.1.4
django-htmlmin==0.8.0
django-imagekit==3.2.7
django-markdown-deux==1.0.5
django-passwords==0.3.4
django-recaptcha==1.0.4
django-secure==1.0.1
django-ses==0.7.0
django-storages-redux==1.3
djangorestframework==3.2.3
email-reply-parser==0.3.0
feedparser==5.2.1
hashids==1.1.0
html5lib==0.9999999
markdown2==2.3.0
MySQL-python==1.2.5
oauthlib==1.0.3
pilkit==1.1.13
Pillow==3.0.0
pytz==2015.6
requests==2.8.1
requests-oauthlib==0.5.0
simplejson==3.8.0
six==1.10.0
sqlparse==0.1.15
